fix: [tag] show confirm dialog before remove tag
as title

Log: 移除标记前弹出确认提示框

Bug: https://pms.uniontech.com/bug-view-139885.html

+++ src/plugins/common/dfmplugin-tag/utils/tagmanager.cpp
@@ -334,7 +334,8 @@ void TagManager::deleteFiles(const QList<QUrl> &urls)
         return;
     }
 
-    deleteTags(tagNames);
+    if (DialogManagerInstance->showDeleteFilesClearTrashDialog(urls) == QDialog::Accepted)
+        deleteTags(tagNames);
 }
 
 bool TagManager::canTagFile(const AbstractFileInfoPointer &fileInfo) const
